Fix some maintainability issues
